+++ BPMobileMessaging/Classes/ConactCenterCommunication/Model/Event/HttpResponseDto/ContactCenterEventTypeDto.swift
@@ -22,6 +22,7 @@ enum ContactCenterEventTypeDto: String, Codable, CaseIterable {
     case chatSessionDisconnect = "chat_session_disconnect"
     case chatSessionEnd = "chat_session_end"
     case chatSessionFile = "chat_session_file"
+    case chatSessionSignaling = "chat_session_signaling"
     case unknown
 
     /// Makes a correspondence between an even type and a particular event Dto JSON codable object
@@ -59,6 +60,8 @@ enum ContactCenterEventTypeDto: String, Codable, CaseIterable {
             return ChatSessionEndDto.self
         case .chatSessionFile:
             return ChatSessionFileDto.self
+        case .chatSessionSignaling:
+            return ChatSessionSignalingDto.self
         case .unknown:
             return UnknownEventDto.self
         }