+++ Example/BPMobileMessaging/WebRTC/ScreenSizeUtil.swift
@@ -1,0 +1,21 @@
+//
+//  ScreenSizeUtil.swift
+//  BPMobileMessaging_Example
+//
+//  Created by Artem Mkrtchyan on 4/24/23.
+//  Copyright © 2023 BrightPattern. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class ScreenSizeUtil {
+    
+    static func width() -> CGFloat {
+        return UIScreen.main.bounds.width
+    }
+    
+    static func height() -> CGFloat {
+        return UIScreen.main.bounds.height
+    }
+}